Added ignoreMemberCheck field
When using LoadFromCollection and passing down your own members list,
inheritance appears to throw an exception. So, giving the ability to
bypass the member check. This appears to be in line since the if
statement where it pulls the properties via GetProperties is nearly
identical to most scenarios when someone passes down a list of
properties.